requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
multidict==6.0.4
packaging==23.1
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.7
pymongo==4.3.3
pytest==7.3.1
python-multipart==0.0.6
python-slugify==8.0.1
sniffio==1.3.0
starlette==0.26.1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.22.0
yarl==1.9.2
